requirements.txt
hidraw-pure
graphitesend